Replace image upload with AWS S3

+++ app/actions.ts
@@ -2,12 +2,21 @@
 
 import { OpenAiRecipe } from "@/types"
 import { PrismaClient, Recipe } from "@prisma/client"
-import { del, put } from "@vercel/blob"
 import { revalidatePath } from "next/cache"
 import { redirect } from "next/navigation"
 import { NextResponse } from "next/server"
+import { S3Client, PutObjectCommand, PutObjectCommandInput, DeleteObjectCommand, DeleteObjectCommandInput } from "@aws-sdk/client-s3"
+import { v4 as uuidv4 } from "uuid"
 import OpenAI from "openai"
 import sharp from "sharp"
+
+const s3Client = new S3Client({
+    credentials: {
+        accessKeyId: process.env.AWS_ACCESS_KEY_ID || "",
+        secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY || ""
+    },
+    region: process.env.AWS_REGION
+})
 
 const prisma = new PrismaClient()
 
@@ -56,16 +65,33 @@ export async function generateImage(recipe: Recipe) {
 
     // compress image
     const uncompressedImage = Buffer.from(image_b64, "base64")
-    const imageFilename = recipe.title.toLowerCase().replace(/ /g, "-") + ".jpeg"
-
     const compressedImage = await sharp(uncompressedImage)
         .jpeg({ mozjpeg: true })
         .toBuffer()
 
-    // upload image to Vercel Blob Storage
-    const blob = await put(imageFilename, compressedImage, {
-      access: 'public',
-    })
+    // upload image to S3 bucket
+    if (!process.env.AWS_BUCKET_NAME || !process.env.AWS_REGION) {
+        throw new Error("AWS bucket config is not set")
+    }
+
+    const bucketName = process.env.AWS_BUCKET_NAME
+    const key = uuidv4() + '.jpeg'
+
+    try {
+        const params: PutObjectCommandInput = {
+            Bucket: bucketName,
+            Body: compressedImage,
+            Key: key,
+            ContentType: 'image/jpeg',
+            ACL: 'public-read',
+        }
+
+        await s3Client.send(new PutObjectCommand(params))
+      } catch (error: any) {
+        return Response.json({ error: error.message })
+      }
+
+      const url = `https://${bucketName}.s3.amazonaws.com/${key}`
 
     // update recipe with image URL
     await prisma.recipe.update({
@@ -73,7 +99,7 @@ export async function generateImage(recipe: Recipe) {
             id: recipe.id
         },
         data: {
-            image: blob.url
+            image: url
         }
     })
 
@@ -88,12 +114,25 @@ export async function deleteRecipe(recipe: Recipe) {
         return NextResponse.json({ error: "You are not authorized to delete this recipe" }, { status: 401 })
     }
 
+    if (!process.env.AWS_BUCKET_NAME || !process.env.AWS_REGION) {
+        throw new Error("AWS bucket config is not set")
+    }
+
+    // delete image from S3 bucket
     try {
-        if (recipe.image) await del(recipe.image)
+        if (recipe.image) {
+            const params: DeleteObjectCommandInput = {
+                Bucket: process.env.AWS_BUCKET_NAME,
+                Key: recipe.image.split('/').pop()
+            }
+
+            await s3Client.send(new DeleteObjectCommand(params))
+        }
     } catch (error) {
         console.log("Failed to delete image", error)
     }
 
+    // delete recipe from database
     try {
         await prisma.recipe.delete({
             where: {
